+++ src/web/RestApi.h
@@ -80,6 +80,7 @@ class RestApi {
             if(path == "html/system")         getHtmlSystem(root);
             else if(path == "html/logout")    getHtmlLogout(root);
             else if(path == "html/save")      getHtmlSave(root);
+            else if(path == "html/chk_save")  getHtmlChkSave(root);
             else if(path == "system")         getSysInfo(root);
             else if(path == "generic")        getGeneric(root);
             else if(path == "reboot")         getReboot(root);
@@ -266,9 +267,19 @@ class RestApi {
 
         void getHtmlSave(JsonObject obj) {
             getGeneric(obj.createNestedObject(F("generic")));
-            obj[F("refresh")] = 2;
-            obj[F("refresh_url")] = "/setup";
-            obj[F("html")] = F("settings succesfully save");
+            obj[F("refresh")] = 1;
+            obj[F("refresh_url")] = F("/chk_save");
+            obj[F("html")] = F("saving settings ...");
+        }
+
+        void getHtmlChkSave(JsonObject obj) {
+            getGeneric(obj.createNestedObject(F("generic")));
+            obj[F("refresh")] = (mApp->getLastSaveSucceed()) ? 10 : 1;
+            obj[F("refresh_url")] = mApp->getSavePending() ? F("/chk_save") : F("/setup");
+            if(mApp->getSavePending())
+                obj[F("html")] = F("saving settings ...");
+            else
+                obj[F("html")] = mApp->getLastSaveSucceed() ?  F("settings succesfully saved") : F("failed saving settings");
         }
 
         void getReboot(JsonObject obj) {